# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Hotel",
        "description": "Documentation de l'API Hotel",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Inscription d'un utilisateur",
                "description": "Crée un nouvel utilisateur",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Alice"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "alice@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "image": {
                                        "description": "Image de profil",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé avec succès"
                    },
                    "422": {
                        "description": "Validation des données échouée"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Connexion utilisateur",
                "description": "Authentifie l'utilisateur et retourne un token JWT",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "alice@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie, retourne le token"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Profil utilisateur",
                "description": "Retourne les informations de l'utilisateur connecté",
                "operationId": "9d9bc556f927575aa492a88c28fe5908",
                "responses": {
                    "200": {
                        "description": "Informations de l'utilisateur"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Déconnexion utilisateur",
                "description": "Invalidate le token JWT de l'utilisateur connecté",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Rafraîchir le token JWT",
                "description": "Génère un nouveau token JWT pour l'utilisateur connecté",
                "operationId": "2c27d311a2876560fcb5a754b556cdc5",
                "responses": {
                    "200": {
                        "description": "Nouveau token généré"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/hotels": {
            "get": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Liste des hôtels",
                "operationId": "1f059609a9a7cdebad01e8494549cc8d",
                "responses": {
                    "200": {
                        "description": "Liste des hôtels",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "city": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Créer un nouvel hôtel",
                "description": "Crée un hôtel pour l'utilisateur connecté (JWT requis)",
                "operationId": "c8f412836761802038a38f7c95ad6c9e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "location",
                                    "rooms",
                                    "price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Hotel Luxe"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Dakar"
                                    },
                                    "rooms": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100.5
                                    },
                                    "image": {
                                        "description": "Image de l'hôtel",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Hôtel créé avec succès"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/hotels/{id}": {
            "get": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Afficher un hôtel",
                "description": "Retourne les détails d’un hôtel (JWT requis)",
                "operationId": "32fa458e73403c0f53352a6162667196",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'hôtel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'hôtel"
                    },
                    "404": {
                        "description": "Hôtel non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Mettre à jour un hôtel",
                "description": "Met à jour les informations d’un hôtel appartenant à l’utilisateur connecté (JWT requis)",
                "operationId": "05f1700e0699e62ac666298252adb116",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'hôtel",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Hotel Royal"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Paris"
                                    },
                                    "rooms": {
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 200
                                    },
                                    "image": {
                                        "description": "Nouvelle image de l'hôtel",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Hôtel mis à jour avec succès"
                    },
                    "403": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Supprimer un hôtel",
                "description": "Supprime un hôtel appartenant à l’utilisateur connecté (JWT requis)",
                "operationId": "d7cc9d4ced4fee467530e2b4017bc6c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'hôtel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hôtel supprimé avec succès"
                    },
                    "403": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Hotel": {
                "title": "Hotel",
                "required": [
                    "name",
                    "location",
                    "rooms",
                    "price"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Hotel Paris"
                    },
                    "location": {
                        "type": "string",
                        "example": "Paris"
                    },
                    "rooms": {
                        "type": "integer",
                        "example": 10
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 100.5
                    },
                    "image": {
                        "type": "string",
                        "example": "hotels/hotel1.jpg",
                        "nullable": true
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Entrer le token JWT précédé de \"Bearer \"",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Hotels",
            "description": "Hotels"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ]
}